style: spark 및 spring package 분리
