+++ src/main/java/chess/controller/spring/MoveController.java
@@ -1,4 +1,4 @@
-package chess.spring.controller;
+package chess.controller.spring;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
