+++ src/main/java/chess/repository/spark/ConnectionManager.java
@@ -1,4 +1,4 @@
-package chess.repository;
+package chess.repository.spark;
 
 import java.sql.Connection;
 import java.sql.DriverManager;
