+++ src/main/java/chess/service/spark/ChessService.java
@@ -1,4 +1,4 @@
-package chess.spring.service;
+package chess.service.spark;
 
 import chess.domain.board.ChessBoard;
 import chess.domain.board.ChessBoardGenerator;
@@ -7,50 +7,49 @@ import chess.domain.history.Histories;
 import chess.domain.piece.TeamType;
 import chess.domain.result.Result;
 import chess.domain.result.Scores;
-import chess.spring.dao.ChessDAO;
-import org.springframework.stereotype.Service;
+import chess.repository.spark.ChessRepository;
 
+import java.sql.SQLException;
 import java.util.Objects;
 
-@Service
 public class ChessService {
     private static ChessBoard CACHED_CHESS_BOARD = null;
 
-    private final ChessDAO chessDAO;
+    private final ChessRepository chessRepository;
 
-    public ChessService(ChessDAO chessDAO) {
-        this.chessDAO = chessDAO;
+    public ChessService(ChessRepository chessRepository) {
+        this.chessRepository = chessRepository;
     }
 
-    public ChessBoard findChessBoard() {
+    public ChessBoard findChessBoard() throws SQLException {
         if (Objects.isNull(CACHED_CHESS_BOARD)) {
             CACHED_CHESS_BOARD = new ChessBoard(ChessBoardGenerator.generateDefaultChessBoard());
-            Histories histories = new Histories(chessDAO.findAllHistories());
+            Histories histories = new Histories(chessRepository.findAllHistories());
             histories.restoreChessBoardAsLatest(CACHED_CHESS_BOARD);
         }
         return CACHED_CHESS_BOARD;
     }
 
-    public TeamType findCurrentTeamType() {
-        Histories histories = new Histories(chessDAO.findAllHistories());
+    public TeamType findCurrentTeamType() throws SQLException {
+        Histories histories = new Histories(chessRepository.findAllHistories());
         return histories.findNextTeamType();
     }
 
-    public void move(String current, String destination, String teamType) {
+    public void move(String current, String destination, String teamType) throws SQLException {
         ChessBoard chessBoard = findChessBoard();
         chessBoard.move(Coordinate.from(current), Coordinate.from(destination), TeamType.valueOf(teamType));
-        chessDAO.insertHistory(current, destination, teamType);
+        chessRepository.insertHistory(current, destination, teamType);
     }
 
-    public Result calculateResult() {
+    public Result calculateResult() throws SQLException {
         ChessBoard chessBoard = findChessBoard();
         Scores scores = chessBoard.calculateScores();
         TeamType winnerTeamType = chessBoard.findWinnerTeam();
         return new Result(scores, winnerTeamType);
     }
 
-    public void deleteAllHistories() {
-        chessDAO.deleteAllHistories();
+    public void deleteAllHistories() throws SQLException {
+        chessRepository.deleteAllHistories();
         CACHED_CHESS_BOARD = null;
     }
 }
