+++ src/main/java/chess/service/spring/ChessService.java
@@ -1,4 +1,4 @@
-package chess.service;
+package chess.service.spring;
 
 import chess.domain.board.ChessBoard;
 import chess.domain.board.ChessBoardGenerator;
@@ -7,49 +7,50 @@ import chess.domain.history.Histories;
 import chess.domain.piece.TeamType;
 import chess.domain.result.Result;
 import chess.domain.result.Scores;
-import chess.repository.ChessRepository;
+import chess.repository.spring.ChessDAO;
+import org.springframework.stereotype.Service;
 
-import java.sql.SQLException;
 import java.util.Objects;
 
+@Service
 public class ChessService {
     private static ChessBoard CACHED_CHESS_BOARD = null;
 
-    private final ChessRepository chessRepository;
+    private final ChessDAO chessDAO;
 
-    public ChessService(ChessRepository chessRepository) {
-        this.chessRepository = chessRepository;
+    public ChessService(ChessDAO chessDAO) {
+        this.chessDAO = chessDAO;
     }
 
-    public ChessBoard findChessBoard() throws SQLException {
+    public ChessBoard findChessBoard() {
         if (Objects.isNull(CACHED_CHESS_BOARD)) {
             CACHED_CHESS_BOARD = new ChessBoard(ChessBoardGenerator.generateDefaultChessBoard());
-            Histories histories = new Histories(chessRepository.findAllHistories());
+            Histories histories = new Histories(chessDAO.findAllHistories());
             histories.restoreChessBoardAsLatest(CACHED_CHESS_BOARD);
         }
         return CACHED_CHESS_BOARD;
     }
 
-    public TeamType findCurrentTeamType() throws SQLException {
-        Histories histories = new Histories(chessRepository.findAllHistories());
+    public TeamType findCurrentTeamType() {
+        Histories histories = new Histories(chessDAO.findAllHistories());
         return histories.findNextTeamType();
     }
 
-    public void move(String current, String destination, String teamType) throws SQLException {
+    public void move(String current, String destination, String teamType) {
         ChessBoard chessBoard = findChessBoard();
         chessBoard.move(Coordinate.from(current), Coordinate.from(destination), TeamType.valueOf(teamType));
-        chessRepository.insertHistory(current, destination, teamType);
+        chessDAO.insertHistory(current, destination, teamType);
     }
 
-    public Result calculateResult() throws SQLException {
+    public Result calculateResult() {
         ChessBoard chessBoard = findChessBoard();
         Scores scores = chessBoard.calculateScores();
         TeamType winnerTeamType = chessBoard.findWinnerTeam();
         return new Result(scores, winnerTeamType);
     }
 
-    public void deleteAllHistories() throws SQLException {
-        chessRepository.deleteAllHistories();
+    public void deleteAllHistories() {
+        chessDAO.deleteAllHistories();
         CACHED_CHESS_BOARD = null;
     }
 }
