--- src/main/java/chess/view/InputView.java
@@ -1,23 +1,0 @@
-package chess.view;
-
-import java.util.Arrays;
-import java.util.List;
-import java.util.Scanner;
-
-public class InputView {
-    private static final Scanner SCANNER = new Scanner(System.in);
-    private static final String DELIMITER = " ";
-
-    private InputView() {
-    }
-
-    public static void printGameStartMessage() {
-        System.out.println("체스 게임을 시작합니다.");
-        System.out.println("게임 시작은 start, 종료는 end 명령을 입력하세요.");
-    }
-
-    public static List<String> inputPlayerCommand() {
-        String[] splitCommand = SCANNER.nextLine().split(DELIMITER);
-        return Arrays.asList(splitCommand);
-    }
-}
