+++ src/main/resources/schema.sql
@@ -1,4 +1,4 @@
-CREATE TABLE IF NOT EXISTS History
+CREATE TABLE IF NOT EXISTS HISTORY
 (
     ID
     INT
